Move calendar to components folder
