+++ libs/core/src/components/calendar/_mixins.scss
