+++ libs/core/src/components/calendar/calendar.trans.styles.scss
